+++ Bundles/ProductPackagingUnit/src/Spryker/Zed/ProductPackagingUnit/Dependency/Facade/ProductPackagingUnitToAvailabilityFacadeBridge.php
@@ -37,24 +37,24 @@ class ProductPackagingUnitToAvailabilityFacadeBridge implements ProductPackaging
 
     /**
      * @param string $sku
-     * @param int $quantity
+     * @param float $quantity
      * @param \Generated\Shared\Transfer\StoreTransfer $storeTransfer
      *
      * @return bool
      */
-    public function isProductSellableForStore(string $sku, int $quantity, StoreTransfer $storeTransfer): bool
+    public function isProductSellableForStore(string $sku, float $quantity, StoreTransfer $storeTransfer): bool
     {
         return $this->availabilityFacade->isProductSellableForStore($sku, $quantity, $storeTransfer);
     }
 
     /**
      * @param string $sku
-     * @param int $quantity
+     * @param float $quantity
      * @param \Generated\Shared\Transfer\StoreTransfer $storeTransfer
      *
      * @return int
      */
-    public function saveProductAvailabilityForStore(string $sku, int $quantity, StoreTransfer $storeTransfer): int
+    public function saveProductAvailabilityForStore(string $sku, float $quantity, StoreTransfer $storeTransfer): int
     {
         return $this->availabilityFacade->saveProductAvailabilityForStore($sku, $quantity, $storeTransfer);
     }
